refactor: AppDirectory
